# docs/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : true,
            "type" : "object"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "details", "message", "traceId" ],
        "type" : "object"
      },
...
      "HandoffRequest" : {
        "properties" : {
          "targetClientId" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "targetClientId" ],
        "type" : "object"
      },
      "HandoffResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "url" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "code", "url" ],
        "type" : "object"
      },
...
      "MagicLinkRequest" : {
        "properties" : {
          "client_id" : {
            "minLength" : 1,
            "type" : "string"
          },
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "purpose" : {
            "enum" : [ "LOGIN", "RESET" ],
            "type" : "string"
          },
          "redirect_uri" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "client_id", "email", "purpose" ],
        "type" : "object"
      },
...
      "PasswordRequest" : {
        "properties" : {
          "current" : {
            "format" : "password",
            "type" : "string",
            "writeOnly" : true
          },
          "new" : {
            "format" : "password",
            "minLength" : 1,
            "type" : "string",
            "writeOnly" : true
          },
          "repeat" : {
            "format" : "password",
            "minLength" : 1,
            "type" : "string",
            "writeOnly" : true
          }
        },
        "required" : [ "new", "repeat" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "AgilityHub Core API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/auth/handoff" : {
      "post" : {
        "description" : "Account token with club context. R-01-13: code lasts 60 seconds and is bound to targetClientId.",
        "operationId" : "handoff",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/HandoffRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HandoffResponse"
                }
              }
            },
            "description" : "Code and destination login URL"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "NO_MEMBERSHIP for the destination role"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Create a one-use code for switching apps",
        "tags" : [ "identity" ]
      }
    },
...
    "/api/v1/me/password" : {
      "put" : {
        "description" : "Account token, never impersonated. R-01-05: current is required only when a password already exists.",
        "operationId" : "password",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Password changed"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "PASSWORD_TOO_SHORT, PASSWORD_MISMATCH, PASSWORD_COMPROMISED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "INVALID_CREDENTIALS for current password"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "FORBIDDEN for impersonated accounts"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Set or change the current account password",
        "tags" : [ "identity" ]
      }
    },
...
    "/auth/magic-link" : {
      "post" : {
        "description" : "ANON. R-01-04: neutral 202 whether the account/membership exists or not; no email enumeration.",
        "operationId" : "magicLink",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MagicLinkRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Request accepted"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "RATE_LIMITED; Retry-After in seconds",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "security" : [ ],
        "summary" : "Request a login or password-reset magic link",
        "tags" : [ "identity" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "https://core.agilitydoghub.com"
  }, {
    "url" : "http://localhost:8080"
  } ]
}